HIA-1431: Refactor auth service and remove AuthoriseConsumerService
