+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/config/AuthorisationConfig.kt
@@ -4,6 +4,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Configuration
 import org.springframework.stereotype.Component
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.ConsumerConfig
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.Role
 
 @Configuration
 @Component
@@ -12,4 +13,5 @@ class AuthorisationConfig(
   var consumers: Map<String, ConsumerConfig?> = emptyMap(),
   var certificateRevocationList: List<String> = emptyList(),
   var defaultConsumerName: String? = null,
+  var roles: Map<String?, Role> = uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.roles,
 )
