+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/extensions/RedactionResponseBodyAdvice.kt
@@ -18,7 +18,7 @@ import uk.gov.justice.digital.hmpps.hmppsintegrationapi.limitedaccess.GetCaseAcc
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.redactionconfig.RedactionPolicy
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.roles
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.redaction.RedactionContext
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.AuthorisationService
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.AuthorisationService
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.telemetry.TelemetryService
 
 @ControllerAdvice
@@ -88,7 +88,7 @@ class RedactionResponseBodyAdvice(
     buildList {
       val consumerRoles = authorisationService.consumers()[subjectDistinguishedName]?.roles.orEmpty()
       consumerRoles.forEach { role ->
-        addAll(roles[role]?.redactionPolicies.orEmpty())
+        addAll(authorisationService.getRole(role)?.redactionPolicies.orEmpty())
       }
     }
 }
