+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/services/internal/AuthorisationService.kt
@@ -1,6 +1,5 @@
-package uk.gov.justice.digital.hmpps.hmppsintegrationapi.services
+package uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal
 
-import org.springframework.context.annotation.Configuration
 import org.springframework.stereotype.Component
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.config.AuthorisationConfig
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.events.enums.IntegrationEventType
@@ -9,16 +8,74 @@ import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.Consum
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.ConsumerFilters
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.Role
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.roles.dsl.MappaCategory
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.RoleService
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.onbehalfof.EntraJwtOboService
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.onbehalfof.OboService
 
-@Configuration
 @Component
 class AuthorisationService(
-  private val roleService: RoleService,
   private val authorisationConfig: AuthorisationConfig,
 ) {
+  fun authorised(
+    consumerName: String,
+    requestedPath: String,
+  ): Boolean =
+    authorisedThroughIncludes(consumerName, requestedPath) ||
+      authorisedThroughRole(consumerName, requestedPath)
+
+  fun matches(
+    path: String,
+    pathTemplate: String,
+  ): Boolean =
+    Regex(
+      normalisePath(pathTemplate),
+    ).matches(path)
+
+  fun doesConsumerHaveIncludesAccess(
+    consumerConfig: ConsumerConfig?,
+    requestedPath: String,
+  ): Boolean {
+    consumerConfig?.permissions()?.forEach {
+      if (matches(requestedPath, it)) {
+        return true
+      }
+    }
+    return false
+  }
+
+  fun doesConsumerHaveRoleAccess(
+    consumerRolesInclude: List<String>,
+    requestPath: String,
+  ): Boolean {
+    consumerRolesInclude.forEach {
+      if (matches(requestPath, it)) {
+        return true
+      }
+    }
+    return false
+  }
+
+  private fun authorisedThroughRole(
+    consumerName: String?,
+    requestedPath: String,
+  ): Boolean {
+    val consumerConfig: ConsumerConfig? = consumers()[consumerName]
+    val consumersRoles = consumerConfig?.roles
+    val rolesInclude =
+      buildList {
+        for (consumerRole in consumersRoles.orEmpty()) {
+          addAll(authorisationConfig.roles[consumerRole]?.permissions.orEmpty())
+        }
+      }
+    val roleResult =
+      doesConsumerHaveRoleAccess(rolesInclude, requestedPath)
+    return roleResult
+  }
+
+  private fun authorisedThroughIncludes(
+    consumerName: String?,
+    requestedPath: String,
+  ) = doesConsumerHaveIncludesAccess(consumers()[consumerName], requestedPath)
+
   /**
    * Returns true if the consumer has access to the endpoint.
    */
@@ -44,7 +101,7 @@ class AuthorisationService(
     val merged = mutableSetOf<String>()
     merged.addAll(authorisationConfig.consumers[consumerName]?.permissions().orEmpty())
     for (roleName in authorisationConfig.consumers[consumerName]?.roles ?: emptyList()) {
-      merged.addAll(roleService.getRoles()[roleName]?.permissions.orEmpty())
+      merged.addAll(authorisationConfig.roles[roleName]?.permissions.orEmpty())
     }
     return merged.toList().sorted()
   }
@@ -82,6 +139,8 @@ class AuthorisationService(
 
   fun defaultConsumerName() = authorisationConfig.defaultConsumerName
 
+  fun getRole(roleName: String) = authorisationConfig.roles[roleName]
+
   /**
    * Returns true if the endpoint matches any of the patterns.
    */
@@ -97,7 +156,7 @@ class AuthorisationService(
     val consumerConfig: ConsumerConfig? = authorisationConfig.consumers[consumerName]
     val roles: List<Role>? =
       consumerConfig?.roles?.mapNotNull {
-        roleService.getRoles()[it]
+        authorisationConfig.roles[it]
       }
     return allFilters(consumerConfig, roles)
   }
