--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/services/internal/RoleService.kt
@@ -1,9 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal
-
-import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.roles
-
-@Component
-class RoleService {
-  fun getRoles() = roles
-}
