+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/util/AuthorisationConfigReader.kt
@@ -5,8 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.config.AuthorisationConfig
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.AuthorisationService
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.RoleService
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.AuthorisationService
 
 /**
  * Class to get a Map containing the Authorisation Config for all consumers
@@ -28,6 +27,6 @@ class AuthorisationConfigReader(
     val configFile = fileManager.readFileFromResourcesFolder("application-$environment.yml")
     val config = yamlMapper.readTree(configFile)
     val authConfig = yamlMapper.treeToValue(config.at("/authorisation"), AuthorisationConfig::class.java)
-    return AuthorisationService(RoleService(), authConfig)
+    return AuthorisationService(authConfig)
   }
 }
