+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/config/ConfigTest.kt
@@ -5,8 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import org.springframework.core.io.ClassPathResource
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.AuthorisationService
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.RoleService
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.AuthorisationService
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.util.AuthorisationConfigReader
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.util.FileManager
 import java.io.File
@@ -33,10 +32,12 @@ abstract class ConfigTest {
     return mapper.convertValue(featureConfig, object : TypeReference<Map<String, Boolean>>() {})
   }
 
-  fun parseAuthorisationConfig(config: String): AuthorisationService {
+  fun parseAuthorisationService(config: String): AuthorisationService {
     val conf = mapper.readValue(config, AuthorisationConfig::class.java)
-    return AuthorisationService(RoleService(), conf)
+    return AuthorisationService(conf)
   }
+
+  fun parseAuthorisationConfig(config: String): AuthorisationConfig = mapper.readValue(config, AuthorisationConfig::class.java)
 
   /**
    * Returns a list of all the configured environments.
