+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/config/WebMvcTestConfiguration.kt
@@ -5,8 +5,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.context.annotation.Bean
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.limitedaccess.GetCaseAccess
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.AuthorisationService
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.RoleService
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.internal.AuthorisationService
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.telemetry.TelemetryService
 
 @TestConfiguration
@@ -21,13 +20,9 @@ class WebMvcTestConfiguration {
 
   @Bean
   @ConditionalOnMissingBean
-  fun roleService(): RoleService = RoleService()
-
-  @Bean
-  @ConditionalOnMissingBean
   fun config(): AuthorisationConfig = AuthorisationConfig()
 
   @Bean
   @ConditionalOnMissingBean
-  fun authorisationService(): AuthorisationService = AuthorisationService(roleService(), config())
+  fun authorisationService(): AuthorisationService = AuthorisationService(config())
 }
